// frontend/src/backend.d.ts
export interface CostumeCharacterAuditionForm {
    age?: bigint;
    name: string;
    musicalSkills: string;
    email: string;
    experience: string;
    referredBy: string;
    performancePreferences: string;
    vocalRange: string;
    physicalLimitations: string;
    operationAgreeStatus: string;
    phone: string;
    scheduleConflicts: string;
    costumePreferences: string;
    characterVoices: string;
    whyAudition: string;
}
// ...
export interface ScareActorAuditionForm {
    age?: bigint;
    specialSkills: string;
    conflictSchedule: string;
    name: string;
    email: string;
    experience: string;
    availability: string;
    referredBy: string;
    whyScaryRole: string;
    preferredScareType: string;
    preferenceOutfitType: string;
    preferedWorkingCondition: string;
    physicalLimitations: string;
    operationAgreeStatus: string;
    phone: string;
    favoriteCharacterType: string;
    previousWork: string;
}
export interface FoodItem {
    theme: FoodTheme;
    name: string;
    description: string;
    itemType: FoodType;
    price: number;
    allergens: Array<string>;
    specialNotes: string;
}
export interface RoomSchedule {
    timeSlots: Array<TimeSlot>;
    roomId: bigint;
}
export interface DanceAuditionForm {
    age?: bigint;
    performanceExperience: string;
    name: string;
    workingConditions: string;
    email: string;
    experience: string;
    availability: string;
    referredBy: string;
    whyDancing: string;
    favoriteDanceType: string;
    danceStyles: string;
    physicalLimitations: string;
    operationAgreeStatus: string;
    phone: string;
    scheduleConflicts: string;
    costumePreferences: string;
    previousWork: string;
}
// ...
export interface HauntedHouseSupervisorAuditionForm {
    age: bigint;
    name: string;
    email: string;
    experience: string;
    availability: string;
    phone: string;
    leadershipExperience: string;
}
// ...
export interface UsherAuditionForm {
    age: bigint;
    name: string;
    email: string;
    experience: string;
    availability: string;
    phone: string;
}